requirements.txt
Flask==3.0.0
rembg==2.0.72
Pillow>=10.0.0
flask-cors==4.0.0
gunicorn==21.2.0
